rename qt_settings -> qt_preferences
