+++ napari/_qt/qt_preferences.py
